Step 1/n for refactoring the iree/compiler/Conversion/* directory. (#6228)
This just moves a few files around. Follow up changes will move the
methods to the right places

Renames CodegenUtils -> Utils
Moves Common/Transforms.* -> Tranforms/Transforms.*
Moves Common/LaunchConfig.* -> LinalgToSPIRV/LaunchConfig.*
Eventually the goal is

Utils folder only has utilities that do not modify IR
Transforms has methods that modify the IR
Common has passes common across all backends.
HLOToLinalg moves out of Conversion/
LinalgToLinalg moves out of Conversion/
VectorToLLVM gets folded into LinalgToLLVM
