+++ iree/compiler/Conversion/Common/BUILD
@@ -37,20 +37,15 @@ cc_library(
         "FlattenMemRefSubspanPass.cpp",
         "FoldTensorExtractOpPass.cpp",
         "ForOpCanonicalizationPass.cpp",
-        "LaunchConfig.cpp",
         "LinalgBufferizePass.cpp",
         "OptimizeVectorTransferPass.cpp",
         "SetNumWorkgroupsPass.cpp",
-        "Transforms.cpp",
-    ],
-    hdrs = [
-        "LaunchConfig.h",
-        "Transforms.h",
     ],
     deps = [
         "//iree/compiler/Conversion:PassHeaders",
-        "//iree/compiler/Conversion/CodegenUtils",
         "//iree/compiler/Conversion/Common:FoldTensorExtractOpIncGen",
+        "//iree/compiler/Conversion/Transforms",
+        "//iree/compiler/Conversion/Utils",
         "//iree/compiler/Dialect/Flow/IR",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/IREE/IR",
