+++ iree/compiler/Conversion/LinalgToLLVM/BUILD
@@ -29,8 +29,9 @@ cc_library(
     ],
     deps = [
         "//iree/compiler/Conversion:PassHeaders",
-        "//iree/compiler/Conversion/CodegenUtils",
         "//iree/compiler/Conversion/Common",
+        "//iree/compiler/Conversion/Transforms",
+        "//iree/compiler/Conversion/Utils",
         "//iree/compiler/Conversion/VectorToLLVM",
         "//iree/compiler/Dialect/Flow/IR",
         "//iree/compiler/Dialect/HAL/IR",
