+++ iree/compiler/Conversion/LinalgToSPIRV/BUILD
@@ -19,6 +19,7 @@ cc_library(
         "ConvertToSPIRVPass.cpp",
         "FoldGPUProcessorIDUses.cpp",
         "KernelDispatchUtils.cpp",
+        "LaunchConfig.cpp",
         "Passes.cpp",
         "TileAndVectorizeInOneWorkgroupPass.cpp",
         "Utils.cpp",
@@ -28,15 +29,17 @@ cc_library(
     ],
     hdrs = [
         "KernelDispatchUtils.h",
+        "LaunchConfig.h",
         "MemorySpace.h",
         "Utils.h",
     ],
     deps = [
         "//iree/compiler/Conversion:PassHeaders",
-        "//iree/compiler/Conversion/CodegenUtils",
         "//iree/compiler/Conversion/Common",
         "//iree/compiler/Conversion/HLOToLinalg",
         "//iree/compiler/Conversion/LinalgToVector",
+        "//iree/compiler/Conversion/Transforms",
+        "//iree/compiler/Conversion/Utils",
         "//iree/compiler/Dialect/Flow/IR",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/HAL/IR:HALDialect",
