+++ iree/compiler/Conversion/LinalgToSPIRV/KernelDispatchUtils.cpp
@@ -15,10 +15,10 @@
 
 #include "iree/compiler/Conversion/LinalgToSPIRV/KernelDispatchUtils.h"
 
-#include "iree/compiler/Conversion/CodegenUtils/FunctionUtils.h"
-#include "iree/compiler/Conversion/Common/LaunchConfig.h"
+#include "iree/compiler/Conversion/LinalgToSPIRV/LaunchConfig.h"
 #include "iree/compiler/Conversion/LinalgToSPIRV/Utils.h"
 #include "iree/compiler/Conversion/Passes.h"
+#include "iree/compiler/Conversion/Utils/Utils.h"
 #include "iree/compiler/Dialect/IREE/IR/IREEOps.h"
 #include "iree/compiler/Dialect/Shape/IR/ShapeOps.h"
 #include "llvm/Support/Debug.h"
