+++ iree/compiler/Conversion/LinalgToSPIRV/Utils.cpp
@@ -12,8 +12,8 @@
 
 #include "iree/compiler/Conversion/LinalgToSPIRV/Utils.h"
 
-#include "iree/compiler/Conversion/CodegenUtils/MarkerUtils.h"
 #include "iree/compiler/Conversion/LinalgToSPIRV/MemorySpace.h"
+#include "iree/compiler/Conversion/Utils/MarkerUtils.h"
 #include "mlir/Dialect/GPU/GPUDialect.h"
 #include "mlir/Dialect/Linalg/IR/LinalgOps.h"
 #include "mlir/Dialect/Linalg/Utils/Utils.h"
