+++ iree/compiler/Conversion/Transforms/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Dialect/HAL/Target/SPIRVCommon/BUILD                           #
+# iree/compiler/Conversion/Transforms/BUILD                                    #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -8,35 +8,26 @@
 # To disable autogeneration for this file entirely, delete this header.        #
 ################################################################################
 
-if(NOT "${IREE_TARGET_BACKEND_VULKAN-SPIRV}" AND
-   NOT "${IREE_TARGET_BACKEND_METAL-SPIRV}")
-  return()
-endif()
-
 iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    SPIRVCommon
+    Transforms
   HDRS
-    "SPIRVTarget.h"
+    "Transforms.h"
   SRCS
-    "SPIRVTarget.cpp"
+    "Transforms.cpp"
   DEPS
     LLVMSupport
+    MLIRAffineUtils
     MLIRIR
-    MLIRSPIRV
-    MLIRSPIRVConversion
-    MLIRSPIRVSerialization
+    MLIRLinalg
+    MLIRMemRef
+    MLIRPass
     MLIRSupport
-    iree::compiler::Conversion::Common
-    iree::compiler::Conversion::LinalgToSPIRV
-    iree::compiler::Conversion::PassHeaders
+    MLIRTransforms
     iree::compiler::Conversion::Utils
-    iree::compiler::Dialect::Flow::IR
-    iree::compiler::Dialect::HAL::Target
-    iree::compiler::Dialect::IREE::IR
-    iree::compiler::Dialect::Shape::IR
+    iree::compiler::Dialect::HAL::IR
   PUBLIC
 )
