+++ iree/compiler/Conversion/Utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Conversion/CodegenUtils/BUILD                                  #
+# iree/compiler/Conversion/Utils/BUILD                                         #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,15 +12,15 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    CodegenUtils
+    Utils
   HDRS
-    "FunctionUtils.h"
     "MarkerUtils.h"
     "TransformUtils.h"
+    "Utils.h"
   SRCS
-    "FunctionUtils.cpp"
     "MarkerUtils.cpp"
     "TransformUtils.cpp"
+    "Utils.cpp"
   DEPS
     LLVMSupport
     MLIRAffine
